refactor(admin)!: fold admin-client into admin-react
First step of the admin-* consolidation in #3902 item 2.

admin-client was 617 lines across five files with exactly one consumer,
admin-react, whose own index already did 'export * from admin-client'. Nobody
could depend on the smaller surface because nothing exposed it separately, so
the split bought a package boundary and a published version line and no
decoupling.

The modules move to admin-react/src/client and are published at
@voyant-travel/admin-react/client. The root export is unchanged, including the
admin-contracts surface the client re-exported - that re-export moved with the
code rather than to the barrel, so importers of the subpath still get
descriptors, types and AdminApiError from one place.

Extension support is untouched. A real extension - smartbill-app - depends on
admin-extension-sdk, ui and apps, and imports none of the admin-* packages
being consolidated. The host side of the extension seam lives in
@voyant-travel/admin, which this does not modify.

Regenerated dep-paths and dist-configs, updated the changesets fixed group,
and corrected the admin-contracts doc comments that named the old package.
