--- packages/admin-client/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@voyant-travel/admin-client",
-  "version": "0.131.1",
-  "license": "Apache-2.0",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "typecheck": "tsc -p tsconfig.typecheck.json",
-    "lint": "biome check src/",
-    "test": "vitest run --passWithNoTests",
-    "build": "tsc -p tsconfig.build.json",
-    "clean": "rm -rf dist tsconfig.tsbuildinfo",
-    "prepack": "pnpm run build"
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public",
-    "exports": {
-      ".": {
-        "types": "./dist/index.d.ts",
-        "import": "./dist/index.js",
-        "default": "./dist/index.js"
-      }
-    },
-    "main": "./dist/index.js",
-    "types": "./dist/index.d.ts"
-  },
-  "dependencies": {
-    "@voyant-travel/admin-contracts": "workspace:^",
-    "zod": "catalog:"
-  },
-  "devDependencies": {
-    "@voyant-travel/voyant-typescript-config": "workspace:^",
-    "typescript": "catalog:",
-    "vitest": "catalog:"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/voyant-travel/voyant.git",
-    "directory": "packages/admin-client"
-  }
-}
