--- packages/admin-client/src/index.ts
@@ -1,6 +1,0 @@
-// Re-export the contract surface so consumers import descriptors, types, and
-// `AdminApiError` from one place.
-export * from "@voyant-travel/admin-contracts"
-export * from "./auth.js"
-export * from "./client.js"
-export * from "./http.js"
