--- packages/admin-client/tsconfig.build.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["./tsconfig.json", "../typescript-config/dep-paths.json"]
-}
