--- packages/admin-client/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@voyant-travel/voyant-typescript-config/base.json",
-  "compilerOptions": {
-    "composite": false,
-    "outDir": "dist",
-    "rootDir": "src",
-    "module": "ESNext",
-    "moduleResolution": "Bundler"
-  },
-  "include": ["src/**/*"]
-}
