--- packages/admin-client/tsconfig.typecheck.json
@@ -1,6 +1,0 @@
-{
-  "extends": ["./tsconfig.json", "../typescript-config/dep-paths.json"],
-  "compilerOptions": {
-    "noEmit": true
-  }
-}
