--- packages/admin-client/vitest.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from "vitest/config"
-
-export default defineConfig({
-  test: {
-    environment: "node",
-    include: ["src/**/*.test.ts"],
-    passWithNoTests: true,
-  },
-})
