+++ packages/admin-react/src/index.ts
@@ -1,6 +1,6 @@
 // Re-export the full admin SDK surface (operation descriptors, client factory,
 // types, and errors) so React consumers import everything from one place.
-export * from "@voyant-travel/admin-client"
+export * from "./client/index.js"
 export * from "./hooks.js"
 export * from "./provider.js"
 export * from "./query-keys.js"
