+++ packages/inventory/tsconfig.build.json
@@ -89,7 +89,6 @@
       "@voyant-travel/admin-app/router": ["../admin-app/dist/router.d.ts"],
       "@voyant-travel/admin-app/runtime": ["../admin-app/dist/runtime.d.ts"],
       "@voyant-travel/admin-app/workspace": ["../admin-app/dist/workspace.d.ts"],
-      "@voyant-travel/admin-client": ["../admin-client/dist/index.d.ts"],
       "@voyant-travel/admin-contracts": ["../admin-contracts/dist/index.d.ts"],
       "@voyant-travel/admin-extension-sdk": ["../admin-extension-sdk/dist/index.d.ts"],
       "@voyant-travel/admin-extension-sdk/client": ["../admin-extension-sdk/dist/client.d.ts"],
@@ -104,6 +103,7 @@
       "@voyant-travel/admin-host/start": ["../admin-host/dist/start.d.ts"],
       "@voyant-travel/admin-host/workspace": ["../admin-host/dist/workspace.d.ts"],
       "@voyant-travel/admin-react": ["../admin-react/dist/index.d.ts"],
+      "@voyant-travel/admin-react/client": ["../admin-react/dist/client/index.d.ts"],
       "@voyant-travel/admin-react/hooks": ["../admin-react/dist/hooks.d.ts"],
       "@voyant-travel/admin-react/provider": ["../admin-react/dist/provider.d.ts"],
       "@voyant-travel/admin-react/query-keys": ["../admin-react/dist/query-keys.d.ts"],
